requirements.txt
streamlit
pymupdf
pillow
pytz